feat: added tier price selection

+++ internal/model/store_model.go
@@ -10,7 +10,8 @@ type Store struct {
 }
 
 type Tier struct {
-	Amount      int64    `json:"amount" bson:"amount"`
+	Amount      float64    `json:"amount" bson:"amount"`
 	Description string   `json:"description" bson:"description"`
 	Benefits    []string `json:"benefits" bson:"benefits"`
 }
+
